requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
chromadb==0.4.18
pandas==2.1.3
sentence-transformers==2.2.2
google-generativeai==0.3.1
python-dotenv==1.0.0
pydantic==2.5.0